backend/pyproject.toml
[project]
name = "teachme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.28.0",
]
